test(comment): extract tests to comment_tests.rs sibling file
The inline `#[cfg(test)] mod tests` block in src/client/comment.rs
fell outside sonar.cpd.exclusions=**/*_tests.rs (which already
covers bug_tests.rs and group_tests.rs siblings). SonarCloud
flagged the wiremock setup boilerplate in two of the new tests
(hybrid_uses_xmlrpc_directly and xmlrpc_mode_skips_rest, ~15
shared lines each — 13.5% duplication on new code, gate is 3%).

Extracting to comment_tests.rs matches the existing convention
and lets the cpd-exclusion rule do its job: test-fixture
boilerplate is categorically different from production
duplication, and forcing extraction would create bad abstractions
per the project's discipline rule.

No test logic changes — pure file move + linker declaration.
